eval2/requirements.txt
langchain==0.3.24
langchain_core==0.3.55
langchain_openai==0.3.14
pydantic==2.11.3
python-dotenv==1.1.0
pandas ==2.0.3
google-generativeai == 0.8.3